backend/requirements.txt
crewai
langchain-ibm
crewai-tools
langchain-openai
python-dotenv
